fix nuovo libro

+++ src/app/nuovo-libro/nuovo-libro.component.ts
@@ -22,7 +22,7 @@ export class NuovoLibroComponent implements OnInit {
   immagine: File; //l'immagine caricata
   //filePath: string;  //conterrà il percorso generato dall'uuid
   //imageRef: any;
-  newLibro: any = {};
+  newLibro: any = {}; //non rendere di classe Libro
   progressoCaricamento = -1;
   uuidv4 = require('uuid/v4');
   idLibroDaURL: string;
@@ -87,20 +87,25 @@ export class NuovoLibroComponent implements OnInit {
   }
 
   submitForm() {
+    console.log("submit")
     this.newLibro.titolo = this.uploadForm.value.titolo;
     this.newLibro.isbn = this.uploadForm.value.isbn;
     this.newLibro.prezzo = this.uploadForm.value.prezzo;
     this.newLibro.descrizione = this.uploadForm.value.descrizione;   //aggiorno newLibro
     this.newLibro.id_utente = this.userService.utente.uid;
-
-    if(this.pathNuovaFoto){
+    /*
+    let data = Date.now();
+    this.newLibro.data = data;
+    */
+    if(this.pathNuovaFoto&&this.newLibro.imagePath){  //se c'è una nuova foto e il libro ne ha una vecchia
       console.log("cancello vecchia foto");
       this.storage.ref(this.newLibro.imagePath).delete();
     }
-    this.newLibro.imagePath = this.pathNuovaFoto || this.newLibro.imagePath;
-    this.newLibro.imageUrl = this.urlNuovaFoto || this.newLibro.imageUrl;
+    this.newLibro.imagePath = this.pathNuovaFoto || this.newLibro.imagePath || "";   //Aggiungo "" per evitare errori con undefined
+    this.newLibro.imageUrl = this.urlNuovaFoto || this.newLibro.imageUrl || "";
     const idLibro = this.newLibro.id || this.uuidv4(); //se il libro non ha un suoi id (quindi è nuovo), glielo imposto con uuid
-    
+    //this.newLibro.id = null;
+    console.log(this.newLibro);
     this.db.collection("books").doc(idLibro).set(this.newLibro)
       .catch(err => console.log(err))
       .then(resolve => {
@@ -124,118 +129,4 @@ export class NuovoLibroComponent implements OnInit {
       }
     })
   }
-  /*
-  ngOnInit() {
-    this.idLibroDaURL = this.route.snapshot.queryParams.id_libro;
-    if (this.idLibroDaURL) {
-      this.stoCercandoLibroDaModificare = true;
-      this.cercaLibro();
-    }
-    this.uploadForm = new FormGroup({
-      "titolo": new FormControl("", Validators.required),
-      "isbn": new FormControl(null, Validators.required),
-      "prezzo": new FormControl("", Validators.required), //inserire pattern moneta
-      "descrizione": new FormControl("", Validators.required)
-    })
-  }
-  */
-  /*
-    submitForm() {
-      //console.log(this.uploadForm);
-      this.newLibro = <Libro>this.uploadForm.value;
-      this.newLibro.id_utente = this.userService.utente.uid;
-      this.newLibro.prezzo = <number>this.newLibro.prezzo;
-      this.newLibro.titolo = this.newLibro.titolo.toLowerCase();    //titolo lowercase altrimenti ricerca catalogo non funziona
-  
-      this.progressoCaricamento = 0; //inizio il caricamento (scompare il form)
-      this.caricaImmagineStorage();   //carico prima l'immagine
-    }
-  
-    annullaForm() {
-      const dialogRef = this.matDialog.open(PerditaModificheComponent,{data: {titolo: "Uscire?", descrizione : "Confermando le modifiche andranno perse"}});
-      dialogRef.afterClosed().subscribe(result => {
-        if (result) {
-          this.router.navigate(["/account"], { queryParams: { inserimentoLibro: 0 } });
-        }
-      })
-    }
-  
-    immagineCaricata(event) {
-      this.immagine = <File>event.target.files[0];
-    }
-  
-    caricaImmagineStorage() {
-      if (this.immagine) {
-        this.filePath = this.uuidv4();    //percorso random
-        let upload = this.storage.upload(this.filePath, this.immagine)
-  
-        upload.percentageChanges().subscribe(percentage => {
-          this.progressoCaricamento = percentage / 3;
-        })
-  
-        upload.then(result => {
-          this.progressoCaricamento = 33;
-          result.ref.getDownloadURL().then(result => {  //una 
-            this.newLibro.imageUrl = result;  //collego donloadUrl a propietà imageUrl del libro
-            //console.log(result);
-            this.caricaOMoficicaLibro();
-  
-          }, err => console.log("errore getDownloadUrl", err))
-        }, err => {
-          console.log("errore caricamento immagine", err);
-          this.router.navigate(["/account"], { queryParams: { inserimentoLibro: 2 } }); //se errore nel caricamento immagine torno a /account
-        })
-  
-      }
-      else {
-        this.caricaOMoficicaLibro();   //se non è caricata un'immagine
-      }
-    }
-  
-    caricaLibro() {
-      this.db.collection("books").add(this.newLibro)  //pubblico il libro
-        .then(val => {
-          //console.log(val);
-          this.progressoCaricamento = 100;
-          setTimeout(() => this.router.navigate(["/account"], { queryParams: { inserimentoLibro: 1 } }), 1000);
-  
-        }, err => {
-          this.router.navigate(["/account"], { queryParams: { inserimentoLibro: 2 } });
-        })
-    }
-  
-    cercaLibro() {   //cerca il libro corrispondente all'url
-      this.db.collection("books").doc(this.idLibroDaURL).valueChanges().subscribe(val => {
-        this.stoCercandoLibroDaModificare = false;
-        const libroCercato = <Libro>val;
-        this.uploadForm.setValue({ "titolo": libroCercato.titolo, "isbn": libroCercato.isbn, "prezzo": libroCercato.prezzo, "descrizione": libroCercato.descrizione })
-      })
-    }
-  
-    modificaLibro() {
-      this.db.collection("books").doc(this.idLibroDaURL).set(this.newLibro).then(
-        result => {
-          //console.log(result);
-          this.progressoCaricamento = 100;
-          setTimeout(() => this.router.navigate(["/account"], { queryParams: { inserimentoLibro: 1 } }), 1000);
-        },
-        err => {
-          //console.log(err),
-          this.router.navigate(["/account"], { queryParams: { inserimentoLibro: 2 } });
-        }
-      )
-    }
-  
-    caricaOMoficicaLibro(){   //qui si decide se il libro verrà aggiunto a db o modificato uno esistente
-      this.progressoCaricamento = 66;
-      if (!this.idLibroDaURL) { //se non sto modificando un libro
-        //console.log("nuovo libro")
-        this.caricaLibro(); //passo al caricamento libro
-      }
-      else {  //se sto modificando un libro
-        //console.log("modifico libro")
-        this.modificaLibro();
-      }
-    }
-    */
 }
